pyproject.toml
[tool.poetry]
name = "salmonbot"
version = "0.1.0"
description = "Underactuated Final Project Spring 2023"
authors = ["Erick Fuentes <erick@csail.mit.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
drake = "^1.14.0"
numpy = "^1.24.2"
matplotlib = "^3.7.1"


[tool.poetry.group.dev.dependencies]
pynvim = "^0.4.3"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
